+++ tests/Xemo.Tests/Cocoon/FirstTests.cs
@@ -1,7 +1,8 @@
 ﻿using Xemo.Cluster;
+using Xemo.Cocoon;
 using Xunit;
 
-namespace Xemo.Cocoon.Tests
+namespace Xemo.Tests.Cocoon
 {
     public sealed class FirstTests
     {
